# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Gestion de Biens Immobiliers - Digitup Company",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api"
        }
    ],
    "paths": {
...
        "/properties": {
            "get": {
                "tags": [
                    "Properties"
                ],
                "summary": "Liste des biens immobiliers",
                "operationId": "8adf02ac3c8ab1b3c0257b9c8879cb50",
                "parameters": [
                    {
                        "name": "city",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Alger"
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "appartement",
                                "villa",
                                "terrain",
                                "bureau",
                                "local_commercial"
                            ],
                            "example": "appartement"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "disponible",
                                "vendu",
                                "location"
                            ],
                            "example": "disponible"
                        }
                    },
                    {
                        "name": "price_min",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "example": 10000000
                        }
                    },
                    {
                        "name": "price_max",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "example": 20000000
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "vue mer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15,
                            "example": 10
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1,
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des biens récupérée avec succès"
                    }
                }
            },
            "post": {
                "tags": [
                    "Properties"
                ],
                "summary": "Créer un nouveau bien",
                "operationId": "dec34a73835ae125a06d8a0fa64e38e6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "type",
                                    "rooms",
                                    "surface",
                                    "price",
                                    "city",
                                    "district",
                                    "description"
                                ],
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "appartement",
                                            "villa",
                                            "terrain",
                                            "bureau",
                                            "local_commercial"
                                        ],
                                        "example": "appartement"
                                    },
                                    "rooms": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "surface": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 85.5
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 15000000
                                    },
                                    "city": {
                                        "type": "string",
                                        "example": "Alger"
                                    },
                                    "district": {
                                        "type": "string",
                                        "example": "Hydra"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Bel appartement avec vue sur mer"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "disponible",
                                            "vendu",
                                            "location"
                                        ],
                                        "example": "disponible"
                                    },
                                    "is_published": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Bien créé avec succès"
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Action non autorisée"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/properties/{id}": {
            "get": {
                "tags": [
                    "Properties"
                ],
                "summary": "Afficher un bien spécifique",
                "operationId": "14f780f6e0125a67285fd2acbfe6bfc3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails du bien récupérés"
                    },
                    "404": {
                        "description": "Bien non trouvé"
                    },
                    "403": {
                        "description": "Accès non autorisé"
                    }
                }
            },
            "put": {
                "tags": [
                    "Properties"
                ],
                "summary": "Mettre à jour un bien",
                "operationId": "8655c8954d57c6ff4e545e5ebdd141bb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "appartement",
                                            "villa",
                                            "terrain",
                                            "bureau",
                                            "local_commercial"
                                        ],
                                        "example": "villa"
                                    },
                                    "rooms": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "surface": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 150
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 25000000
                                    },
                                    "city": {
                                        "type": "string",
                                        "example": "Oran"
                                    },
                                    "district": {
                                        "type": "string",
                                        "example": "Les Planteurs"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Villa rénovée avec jardin"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "disponible",
                                            "vendu",
                                            "location"
                                        ],
                                        "example": "disponible"
                                    },
                                    "is_published": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Bien mis à jour avec succès"
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Action non autorisée"
                    },
                    "404": {
                        "description": "Bien non trouvé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Properties"
                ],
                "summary": "Supprimer un bien",
                "operationId": "999e4323eea15c4e2751dcdd6642b9c5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Bien supprimé avec succès"
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Action non autorisée"
                    },
                    "404": {
                        "description": "Bien non trouvé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/properties/{id}/toggle-publish": {
            "post": {
                "tags": [
                    "Properties"
                ],
                "summary": "Basculer le statut de publication",
                "operationId": "70e60395ab0af4eecc460dcc6978841b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Statut modifié avec succès"
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Action non autorisée"
                    },
                    "404": {
                        "description": "Bien non trouvé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}